cleanup: reorganize a bit
